requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
potrace>=0.1.3
rich>=13.0.0
click>=8.1.0
